fixing precipitation color scheme
